+++ src/internal/data.ts
@@ -1,14 +1,12 @@
 import type { Model, SchemaDefinition } from 'mongoose';
 
-import type { IGlobalOptions, VirtualOptions } from '../types';
+import type { IGlobalOptions } from '../types';
 import { Severity } from './constants';
 
 /** Schema Map */
 export const schemas: Map<string, SchemaDefinition> = new Map();
 /** Models Map */
 export const models: Map<string, Model<any>> = new Map();
-/** Virtuals Map */
-export const virtuals: Map<string, Map<string, VirtualOptions>> = new Map();
 /** Constructors Map */
 export const constructors: Map<string, NewableFunction> = new Map();
 /** Global Options */